+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- [`getMyFormsApp()`](./docs/forms-apps.md#getmyformsapp) function
+
 ## 0.2.8 (2019-05-23)
 
 ### Added