+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [`getMyFormsApp()`](./docs/forms-apps.md#getmyformsapp) function
+- `injectForms` parameter to `getForm()` function
 
 ## 0.2.8 (2019-05-23)